dev: refactor core-profiler - moved tracks out
