+++ plugins/woocommerce/changelog/dev-refactor-core-profiler-tracks
@@ -1,0 +1,4 @@
+Significance: patch
+Type: dev
+
+Tidied up core profiler's tracks actions
